pyproject.toml
[project]
name = "augmentai"
version = "0.1.0"
description = "LLM-powered data augmentation policy designer with domain-safe, task-aware policies"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "AugmentAI Contributors"}
]
keywords = ["augmentation", "data-augmentation", "llm", "machine-learning", "computer-vision"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Image Processing",
]

dependencies = [
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "rich>=13.0",
    "typer>=0.9",
    "openai>=1.0",
    "httpx>=0.25",
    "albumentations>=1.3",
    "numpy>=1.24",
    "pillow>=10.0",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
kornia = ["kornia>=0.7", "torch>=2.0"]
torchvision = ["torchvision>=0.16", "torch>=2.0"]
all = ["kornia>=0.7", "torchvision>=0.16", "torch>=2.0"]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=23.0",
    "ruff>=0.1",
    "mypy>=1.0",
]

[project.scripts]
augmentai = "augmentai.cli.app:main"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["augmentai*"]

[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311", "py312"]

[tool.ruff]
line-length = 100
target-version = "py39"
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
